Fixes for Windows compatibility
- getline is not part of libc on Windows, so brought in open source version from NetBSD
- g++ in Windows complains about long vs int when setting coefficients
- Need to bring forward linking directories for gmp and mpfr as they appear after PKG_LIBS in linker command by default
- Add creation of config.h which is handled by configure script on *nix

+++ src/FandV_keys.cpp
@@ -11,6 +11,7 @@ using namespace RcppParallel;
 
 #include <limits.h>
 #include <string>
+#include "getline.h"
 
 #include "FandV_keys.h"
 #include "FandV_ct.h"
@@ -37,9 +38,9 @@ void FandV_pk::enc(int m, FandV_ct& ct) const {
   
   // Random numbers
   for(int i=0; i<p.Phi.length()-1; i++) {
-    u.set_coeff(i, lround(R::rnorm(0.0,p.sigma))); // u
-    ct.c0.set_coeff(i, lround(R::rnorm(0.0,p.sigma))); // e1
-    ct.c1.set_coeff(i, lround(R::rnorm(0.0,p.sigma))); // e2
+    u.set_coeff(i, (int) lround(R::rnorm(0.0,p.sigma))); // u
+    ct.c0.set_coeff(i, (int) lround(R::rnorm(0.0,p.sigma))); // e1
+    ct.c1.set_coeff(i, (int) lround(R::rnorm(0.0,p.sigma))); // e2
   }
   
   // Binary conversion of message
@@ -68,9 +69,9 @@ void FandV_pk::encbinary(IntegerVector m, FandV_ct& ct) const {
   
   // Random numbers
   for(int i=0; i<p.Phi.length()-1; i++) {
-    u.set_coeff(i, lround(R::rnorm(0.0,p.sigma))); // u
-    ct.c0.set_coeff(i, lround(R::rnorm(0.0,p.sigma))); // e1
-    ct.c1.set_coeff(i, lround(R::rnorm(0.0,p.sigma))); // e2
+    u.set_coeff(i, (int) lround(R::rnorm(0.0,p.sigma))); // u
+    ct.c0.set_coeff(i, (int) lround(R::rnorm(0.0,p.sigma))); // e1
+    ct.c1.set_coeff(i, (int) lround(R::rnorm(0.0,p.sigma))); // e2
   }
   
   // Binary conversion of message
